requirements.txt
discords
asyncio
commands
sys
os